refactor(rector): apply level 2 (#8617)
* refactor(rector): set rector level to 2

* refactor(rector): apply rector level 2

+++ interface/modules/zend_modules/module/Carecoordination/src/Carecoordination/Model/CcdaServiceDocumentRequestor.php
@@ -67,7 +67,7 @@ class CcdaServiceDocumentRequestor
                 }
                 sleep(5); // give cpu a rest
                 // now try to connect to the server
-                $result = socket_connect($socket, "127.0.0.1", (int)6661);
+                $result = socket_connect($socket, "127.0.0.1", 6661);
                 if ($result === false) {
                     $errorCode = socket_last_error($socket);
                     $errorMsg = socket_strerror($errorCode);
